feat: restructure layout components and update import paths for auth and dashboard layouts
