+++ src/routes/app-routes.tsx
@@ -3,7 +3,7 @@ import { Route, Routes } from "react-router";
 import DashboardRoutes from "./sections/dashboard";
 import RegisterPage from "@/Page/auth/register-page";
 import RecoveryPage from "@/Page/auth/recovery-page";
-import AuthLayout from "@/layouts/auth-layout";
+import AuthLayout from "@/layouts/auth/auth-layout";
 
 export const AppRouter = () => {
   return (
